bug(react-router): fix react router  - refactor naming convention  [Fixes #160577434]
