+++ src/App.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import HomePage from './components/HomePage';
-import LoginPage from './components/LoginPage';
+import HomePage from './components/HomeComponent';
+import LoginComponent from './components/Auth/LoginComponent';
 import './App.scss';
 
 const App = () => (
   <Switch>
     <Route path="/" exact component={HomePage} />
-    <Route path="/login" exact component={LoginPage} />
+    <Route path="/login" exact component={LoginComponent} />
   </Switch>
 );
